Map ordering encoded via generic types
A new abstract type, Order, and two descendeants (RingOrder,
NestedOrder) are now used to encode the ordering of a Map. This will
allow to leave order checks to the Julia compiler.

Alm constructor has been renamed to readAlmFromFITS, for consistency
with readMapFromFITS. The implementation of "conformables" has been
improved, and tests have been added.